+++ app/src/test/java/woowacourse/movie/presenter/reservationDetails/ReservationDetailPresenterTest.kt
@@ -7,7 +7,6 @@ import io.mockk.mockk
 import io.mockk.verify
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.movie.model.reservation.ReservationDao
 
 class ReservationDetailPresenterTest {
     private lateinit var view: ReservationDetailContracts.View
