+++ app/src/test/java/woowacourse/movie/presenter/seatSelection/SeatSelectionPresenterTest.kt
@@ -9,10 +9,10 @@ import io.mockk.mockk
 import io.mockk.verify
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.movie.model.reservation.ReservationDao
 import woowacourse.movie.presenter.MOVIE_TO_RESERVE
 import woowacourse.movie.presenter.SEAT_2_2
 import woowacourse.movie.presenter.SEAT_2_3
+import woowacourse.movie.presenter.reservationDetails.ReservationDao
 
 class SeatSelectionPresenterTest {
     private lateinit var presenter: SeatSelectionPresenter
